+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 4.15.x
 
+* Fix crash when calling `NativeInterface.clearTab()` (from an integration
+  library)
+  [#582](https://github.com/bugsnag/bugsnag-android/pull/582)
 * Fix abort() in native code when storing breadcrumbs with null values in
   metadata
   [#510](https://github.com/bugsnag/bugsnag-android/pull/510)